Moved "*Info" data logic into its own folder for modding.
