+++ game/src/main/java/data/GameSave.java
@@ -25,6 +25,9 @@ import abstracts.Character.GenderType;
 import abstracts.Item;
 import abstracts.SubMenu;
 import data.chunk.SaveChunk;
+import data.mod.AreaInfo;
+import data.mod.HeaderInfo;
+import data.mod.PlayerInfo;
 import entity.Player;
 import level.Area;
 import level.OverWorld;
